First commit: public release
I've been developing this library for a little while to help with
another one of my Myo projects. This is my first attempt at cleaning it
up and publishing it to the world.

I haven't actually tested if this library works yet, but it compiles
correctly and I haven't changed any of the implementation from where
I've been using it in my project successfully.